lint: suggested changes
