+++ genesis-processing/processing/accounts.py
@@ -1,6 +1,8 @@
-from psycopg import Connection
-from db.table_manager import TableManager, DBTypes
 from typing import List
+
+from db.table_manager import DBTypes, TableManager
+from psycopg import Connection
+
 from utils.loggers import get_logger
 
 _logger = get_logger(__name__)
@@ -10,8 +12,7 @@ CHAIN_ID = "chain_id"
 TABLE_ID = "accounts"
 
 
-class AccountsManager():
-
+class AccountsManager:
     def __init__(self, db_conn: Connection):
         columns = (
             (ID, DBTypes.text),
@@ -29,7 +30,9 @@ class AccountsManager():
         accounts_data = self._get_account_data(genesis_data)
         db_accounts = self.table_manager.select_query([ID])
 
-        genesis_accounts_filtered = self._filter_genesis_accounts(accounts_data, db_accounts)
+        genesis_accounts_filtered = self._filter_genesis_accounts(
+            accounts_data, db_accounts
+        )
         with self.table_manager.db_copy() as copy:
             for account in genesis_accounts_filtered:
                 copy.write_row([self._get_account_address(account), chain_id])
@@ -40,7 +43,9 @@ class AccountsManager():
     def _get_account_address(self, account: dict) -> str:
         return str(account["address"])
 
-    def _filter_genesis_accounts(self, accounts_data: List[dict], db_accounts: List[str]) -> List[dict]:
+    def _filter_genesis_accounts(
+        self, accounts_data: List[dict], db_accounts: List[str]
+    ) -> List[dict]:
         """
         Filter out genesis_accounts IDs from accounts_data
 
@@ -50,7 +55,11 @@ class AccountsManager():
         """
         genesis_accounts = [self._get_account_address(x) for x in accounts_data]
 
-        matches = [match for match in set(db_accounts) & set(genesis_accounts)]  # list already indexed accounts
-        genesis_accounts_filtered = filter(lambda account: self._get_account_address(account) not in matches,
-                                           accounts_data)
+        matches = [
+            match for match in set(db_accounts) & set(genesis_accounts)
+        ]  # list already indexed accounts
+        genesis_accounts_filtered = filter(
+            lambda account: self._get_account_address(account) not in matches,
+            accounts_data,
+        )
         return list(genesis_accounts_filtered)
